requirements/airflow_requirements.txt
black==23.3
dbt-core==1.5.0
dbt-extractor==0.4.1
dbt-postgres==1.5.0
Fiona==1.9.3
GeoAlchemy2==0.13.3
geopandas==0.13.0
great-expectations==0.16.14
jupyterlab
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
networkx==2.8.8
pyproj==3.5.0
pytest==7.3.1
pytest-mock==3.10.0
Rtree==1.0.1
scikit-learn==1.2.2
scipy==1.10.1
shapely==2.0.1